Unify public API under Kino.JS.Live and move server implementation to Kino.JS.LiveServer
